backend/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
geopy==2.3.0